Re-write to start simplets branch
The key difference with this branch is that the Data type contains no information about the Node or NodeStorage type contained within. This will require more runtime checks; however, this should be faster to code-up and require less type system gymnastics. This is part of an analysis of the trade-off between runtime and compiletime checks (with associated verbosity considerations).