Fix subtle major bug in aggregate copy methods
Recall: Copy methods, by contract, must return an unbound, precise copy.
-> unbound means, the new version is loosely bound to a SPEC node, not whatever else
-> precise means the node storage and direction (and any other future metadata stored by SPEC)
    is copied over

The old version of copy for Aggregates would just 'create a new one' and hope that was the same.
However, if this would only capture information present in the constructor and not capture actual
changes to the node. (e.g. If Input, Output, or Flipped was called on the data, the broken copied
version would have the data default direction and not the direction as the old one).

HardwareTuple and Vec have been fixed to follow the contracts.
-> Vec was easily fixed by using elemType appropriately

-> HardwareTuple was fixed with a slightly more involved system. The first is that it uses a SimpleCopy trait.
The user (or, the @bundle macro) provides a simplecopy method that simply returns a skeletal copy of the version.
So, all the subfields will be structurally the same. Then, the copy method can call simplecopy and then mimic on
on the skeletal copy to have recursively dig in and copy over all the new metadata
-> mimic is defined at the Data level and tells a Data to mimic the underlying node type as the model. If asSPEC
  is true then 'this' will have nodes SPEC but with metadata (e.g. direction if applicable, width) matching the
  model. If asSPEC is false then the nodes will just be equal (since nodes are all case classes, this is acceptable).
  -> The asSPEC=true case is used for making unbound copies
  -> The asSPEC=false case is used for ensuring a Vecs elements precisely match the elemType
  Note, Element.mimic is what actually implements the node matching. Aggregate.mimics (vec, HWT) just drill down to
  Elements. (Similar mechanism as rebind.)